+++ changelog.md
@@ -4,12 +4,20 @@ API changes / Backward compatible features / Backward compatible bug fixes.
 You do not need to update the version for refactors or code improvements that
 do not change functionality nor do you need to when you you write tests.
 
-## [1.6.3] - 2026-04-07
+## [1.7.1] - 2026-04-07
 
 ### Added
 
 - SnapFit MVC component that snaps blocks together via snap points
 
+## [1.7.0] - 2026-04-07
+
+### Added
+
+- Multiple levels for occipital lobe game
+- Scoring system for occipital lobe game
+- Tutorial mode for occipital lobe game 
+ 
 ## [1.6.2] - 2026-04-07
 
 ### Added